+++ ui/src/explorer_main.tsx
@@ -2,10 +2,10 @@ import ReactDOM from "react-dom/client";
 import React from "react";
 
 import "./tool/wasm_exec.js"
-import {App} from "./explorer/app.tsx";
+import {Explorer} from "./explorer/Explorer.tsx";
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
-        <App/>
+        <Explorer/>
     </React.StrictMode>,
 )
